backend/requirements.txt
fastapi
uvicorn
ollama
pdfplumber
python-multipart
python-dotenv
google-generativeai
